create rename component
